Target handling bugfix
Cache entries are not reloaded by default after an add_subdirectory.
This meant that if a target had multiple tests, only the last one
had correct dependencies, others were ignored.

This is now fixed by reloading the cache variable from the cache.

There was also a bug which prevented the inclusion of "-" in
scope names, also fixed now.